Create a new standalone unit test for the debug device.
