+++ tests/unit/CMakeLists.txt
@@ -79,6 +79,7 @@ set(UNIT_TESTS_SRCS
   fem/test_transfer.cpp
   fem/test_var_order.cpp
   # The following are tested separately (keep the comment as a reminder):
+  # miniapps/test_debug_device.cpp
   # miniapps/test_sedov.cpp
   # miniapps/test_tmop_pa.cpp
   # ceed/test_ceed.cpp
@@ -304,3 +305,20 @@ if (MFEM_USE_MPI)
    add_parallel_miniapp_test(sedov TRUE)
    add_parallel_miniapp_test(tmop_pa FALSE)
 endif(MFEM_USE_MPI)
+
+#-----------------------------------------------------------
+# SERIAL DEBUG-DEVICE TESTS:
+#   debug_device_tests
+#-----------------------------------------------------------
+set(DEBUG_DEVICE_SRCS miniapps/test_debug_device.cpp)
+if (MFEM_USE_CUDA)
+   set_property(SOURCE ${DEBUG_DEVICE_SRCS} PROPERTY LANGUAGE CUDA)
+endif()
+if (MFEM_USE_HIP)
+   set_property(SOURCE ${DEBUG_DEVICE_SRCS}
+                PROPERTY HIP_SOURCE_PROPERTY_FORMAT TRUE)
+endif()
+mfem_add_executable(debug_device_tests unit_test_main.cpp ${DEBUG_DEVICE_SRCS})
+target_link_libraries(debug_device_tests mfem)
+add_dependencies(${MFEM_ALL_TESTS_TARGET_NAME} debug_device_tests)
+add_test(NAME debug_device_tests COMMAND debug_device_tests)
